CrewAIAgents/requirements.txt
crewai
crewai_tools
python-dotenv
#youtube-transcript-api
yt_dlp==2023.11.14
youtube-transcript-api==0.6.1